requirements.txt
cantools
json2yaml